gas analyser: support for refund inputs
n.b. that the input is now expected to be a JSON array of raw gas
kasts, and the output will always be in array form.

With stratification (the default), the "constructors" field (which
used to be called "constructor") will be an array of strings like
["#G0", "#G1", ..., "#Gn"], and should be interpreted as a list of
lower bounds for the initial gas, i.e. should translate into a
condition like (VGas >=Int #G0) andBool ... andBool (VGas >=Int #Gn).

With the flag --no-stratify, the output will be an array of strings,
corresponding to the unstratified gas expressions.