gas analyser: add support for unknowns in gas expressions
basic support was accomplished by just making everything more generic
and introducing the IntOrBlob type, and parsing into the resulting
GasExpr type.

solving support was made possible by adding "extirpation", which is a
way of massaging the gas expression, and solveAlgebraically, which
trivially solves (VGas - X >= 0) => (VGas >= X). extirpation can be
extended further (see the commented out rules, WIP) to add support for
expressions that contain unknowns in addition to multiple occurences
of VGas: in this case we can try to first extirpate all of the
unknowns, and then solve what is left numerically, and recombine with
the algebraic solution that has the unknowns in the end.