kast.js: more accurate pretty printing of modInt
The `%Int` and `modInt` operators have distinct and different meanings within `K`.

The evm-semantics uses `modInt` almost exclusively and the prescence of
`%Int` within a spec is probably an error. This change makes it less
likely that someone will copy and paste a `%` from the debug view into a
spec and use it there, it also brings the output in the debug view more
inline with the spec language.