libexec/klab-report: stop checking for empty files
The call to `read()` was reading the whole file in memory, which would
explode at runtime for very long logs.

This commit removes the check on empty files, which could result in
links being shown in the report page that are effectively useless.

The ideal solution would be to use something like Node's Stream class,
to perform the same check without loading everything in memory. Decided
not to implement this for now, as the issue seems to be relatively
harmless.